requirements.txt
asgiref==3.2.7
click==7.1.2
distro==1.5.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.6
numpy==1.18.3
pandas==1.0.3
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
tabula-py==2.1.0
tqdm==4.45.0
Werkzeug==1.0.1
